requirements.txt
python-telegram-bot
yfinance
numpy
pandas
requests
tqdm
pygsheets
tinvest
plotly
kaleido
pydantic
